Fix bug so cancel when editing an existing match no longer deletes the match

+++ FRC_Scout/app/src/main/java/com/frc2135/android/frc_scout/PreMatchActivity.java
@@ -22,18 +22,18 @@ import org.json.JSONException;
 
 public class PreMatchActivity extends AppCompatActivity
 {
-    private String m_teamIndexStr;
-    private EditText m_competitionField;
-    private AutoCompleteTextView m_scoutNameField;
-    private AutoCompleteTextView m_teamNumberField;
-    private EditText m_matchNumberField;
-    private TextView m_missingFieldErrMsg;
-    private MatchData m_matchData;
-    private CompetitionInfo m_compInfo;
-
     public static final String TAG = "PreMatchActivity";
 
-    private static Scouter m_Scouter;
+    private String               m_teamIndexStr;
+    private EditText             m_competitionField;
+    private AutoCompleteTextView m_scoutNameField;
+    private AutoCompleteTextView m_teamNumberField;
+    private EditText             m_matchNumberField;
+    private TextView             m_missingFieldErrMsg;
+    private MatchData            m_matchData;
+    private CompetitionInfo      m_compInfo;
+    private static Scouter       m_Scouter;
+    private String               m_inEdit; 
 
     private void setTeamNumFromMatchNum()
     {
@@ -64,6 +64,7 @@ public class PreMatchActivity extends AppCompatActivity
         m_Scouter = Scouter.get(getApplicationContext());
 
         String matchId = getIntent().getStringExtra("match_ID");
+        m_inEdit = getIntent().getStringExtra("in_edit");  
         m_matchData = MatchHistory.get(getApplicationContext()).getMatch(matchId);
         m_compInfo = CompetitionInfo.get(getApplicationContext(), m_matchData.getEventCode().trim(), false);
 
@@ -259,9 +260,12 @@ public class PreMatchActivity extends AppCompatActivity
             @Override
             public void onClick(View view)
             {
-                MatchHistory.get(getApplicationContext()).deleteMatch(m_matchData);
-                Intent intent2 = new Intent(PreMatchActivity.this, MatchListActivity.class);
-                startActivity(intent2);
+               // Only delete match if not currently being edited  
+               if (!m_inEdit.equals("yes")) {
+                  MatchHistory.get(getApplicationContext()).deleteMatch(m_matchData);
+               }
+               Intent intent2 = new Intent(PreMatchActivity.this, MatchListActivity.class);
+               startActivity(intent2);
             }
         });
     }
